move pg queries to child branch
