+++ src/bot/actions/index.js
@@ -1,45 +1,13 @@
 const fs = require('fs');
 const util = require('util');
-const { createMessageMetric } = require('../../db/queries/message_metrics');
-const { createUserChat } = require('../../db/queries/user_chats');
-const { createUser } = require('../../db/queries/users');
-const { createChat } = require('../../db/queries/chats');
 const { statsArray } = require('../statistics');
-const { dbClient } = require('../../db/dbClientPg');
+const { dbClient } = require('../../dbClient');
 
 const isCommand = ({ entities }) => !!entities && entities.some((entity) => entity.type === 'bot_command');
 
 const renderMessage = (statsStringsArray) => statsStringsArray
   .filter((statString) => statString.length > 0)
   .join('\n');
-
-const fetchMessageMetrics = ({
-  from, chat, message_id: messageId, date,
-}) => {
-  const asd = 1;
-  return {
-    user: {
-      tg_id: from.id,
-      first_name: from.first_name,
-      last_name: from.last_name,
-    },
-    chat: {
-      tg_id: chat.id,
-      name: chat.title,
-    },
-    messageMetrics: {
-      tg_id: messageId,
-      timestamp: date,
-      photoCount: 0, // определяем наличие фото в сообщении как в contentSuppier
-      videoCount: 0, // определяем наличие видео в сообщении как в contentSuppier
-      questionCount: 0, // если в тексте есть вопрос, ставим 1 (осторожно urlы)
-      stickerSetName: 0,
-      textLength: 0,
-      voiceCount: 0, // ставим 1 если сообщение это голосуха
-      lolReplyForUser: 2, // humoristStat ищем в сообщениии реакцию смеха, и если она есть записываем сюда id юзера из реплая reply_to_message.from.id
-    },
-  };
-};
 
 const stats = async (bot, message) => {
   const statsText = await Promise.all(statsArray.map(async ({ render, collect }) => {
@@ -88,35 +56,9 @@ const onMessage = async (bot, message) => {
       await version(bot, message);
     }
   } else {
-    // message_metrics
-    // users
-    // chats
-
-    console.log('message: ', message);
-
-    const { chat, user, messageMetrics } = fetchMessageMetrics(message);
-
-    const client = await dbClient.connect();
-    try {
-      await client.query('BEGIN');
-      const chatResult = await createChat(client, chat);
-      console.log('chatResult: ', chatResult);
-      const userResult = await createUser(client, user);
-      console.log('userResult: ', userResult);
-      const userChatResult = await createUserChat(client, { user: userResult, chat: chatResult });
-      console.log('userChatResult: ', userChatResult);
-      const messageMetricResult = await createMessageMetric(client, messageMetrics, userChatResult);
-      console.log('messageMetricResult: ', messageMetricResult);
-      await client.query('COMMIT');
-    } catch (e) {
-      await client.query('ROLLBACK');
-      throw e;
-    } finally {
-      client.release();
-    }
-    // await dbClient.queryMessages(async (messages) => {
-    //   await messages.insertOne(message);
-    // });
+    await dbClient.queryMessages(async (messages) => {
+      await messages.insertOne(message);
+    });
   }
 };
